+++ CHANGELOG.md
@@ -5,10 +5,13 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 
 ## [Unreleased]
 ### Added
+<<<<<<< HEAD
 - New runtime Satori client options for time-based caching.
 
 ### Changed
 - Adjust unit of measure for Satori client timeout.
+- Add runtime Satori client additional filters for live-events listing.
+- Add runtime Satori client live-event join function.
 
 ### Fixed
 - Fix migration script to correctly revert console fine-grained access control changes.